tools/formula_ingest/requirements.txt
PyYAML>=6.0.2
mwparserfromhell>=0.6.6
requests>=2.32.3
